// chrome/browser/resources/settings/chromeos/multidevice_page/multidevice_constants.ts
export enum MultiDeviceSettingsMode {
  NO_ELIGIBLE_HOSTS = 0,
  NO_HOST_SET = 1,
  HOST_SET_WAITING_FOR_SERVER = 2,
  HOST_SET_WAITING_FOR_VERIFICATION = 3,
  HOST_SET_VERIFIED = 4,
}

/**
 * Enum of MultiDevice features. Note that this is copied from (and must
 * include an analog of all values in) the Feature enum in
 * //chromeos/ash/services/multidevice_setup/public/mojom/multidevice_setup.mojom.
 */
export enum MultiDeviceFeature {
  BETTER_TOGETHER_SUITE = 0,
  INSTANT_TETHERING = 1,
  MESSAGES = 2,
  SMART_LOCK = 3,
  PHONE_HUB = 4,
  PHONE_HUB_NOTIFICATIONS = 5,
  PHONE_HUB_TASK_CONTINUATION = 6,
  WIFI_SYNC = 7,
  ECHE = 8,
  PHONE_HUB_CAMERA_ROLL = 9,
}

/**
 * Possible states of MultiDevice features. Note that this is copied from (and
 * must include an analog of all values in) the FeatureState enum in
 * //chromeos/ash/services/multidevice_setup/public/mojom/multidevice_setup.mojom.
 */
export enum MultiDeviceFeatureState {
  PROHIBITED_BY_POLICY = 0,
  DISABLED_BY_USER = 1,
  ENABLED_BY_USER = 2,
  NOT_SUPPORTED_BY_CHROMEBOOK = 3,
  NOT_SUPPORTED_BY_PHONE = 4,
  UNAVAILABLE_NO_VERIFIED_HOST = 5,
  UNAVAILABLE_INSUFFICIENT_SECURITY = 6,
  UNAVAILABLE_SUITE_DISABLED = 7,
  FURTHER_SETUP_REQUIRED = 8,
  UNAVAILABLE_TOP_LEVEL_FEATURE_DISABLED = 9,
  UNAVAILABLE_NO_VERIFIED_HOST_CLIENT_NOT_READY = 10,
  UNAVAILABLE_NO_VERIFIED_HOST_NO_ELIGIBLE_HOST = 11,
}
// ...
export enum PhoneHubFeatureAccessProhibitedReason {
  UNKNOWN = 0,
  WORK_PROFILE = 1,
  DISABLED_BY_PHONE_POLICY = 2,
}
// ...
export enum PhoneHubPermissionsSetupFeatureCombination {
  NONE = 0,
  NOTIFICATION = 1,
  MESSAGING_APP = 2,
  CAMERA_ROLL = 3,
  NOTIFICATION_AND_MESSAGING_APP = 4,
  NOTIFICATION_AND_CAMERA_ROLL = 5,
  MESSAGING_APP_AND_CAMERA_ROLL = 6,
  ALL_PERMISSONS = 7,
}